Add files via upload
We will be using the famous Iris Flower dataset. The Iris flower dataset is a multivariate dataset introduced by sir Ronald FIsher in the 1936 as an example of discriminant analysis. The data set consists of 50 samples from each of three species of Iris(Iris setosa,Iris virginica and iris versicolor),so 150 samples in total.Four samples were measured from each sample :the length and the width of the sepals and petals,in centimeters.